requirements.txt
python-dotenv==1.1.0
aiogram==3.20.0.post0